Unify on int types. All bytes and shorts are converted to ints.
All casting removed. Booleans are still used for the Keyboard class.

Memory class:
* File reading can only be done into a byte array, so do that and copy
  values to an int array.
* All store methods mask data with 0xFF to keep only bytes in int array.
* All addresses are masked with 0xFFF for 12-bit addresses.
* All instructions are masked with 0xFFFF for 16-bit instructions.

Registers:
* All registers are ints, but are masked to ensure correct bit length.
* PC and I are masked with 0xFFF for 12-bit addresses.
* SP, delayTimer, and soundTimer are masked with 0xFF for 8-bit values.
* Vx registers are an int array masked with 0xFF for 8-bit values.
* Stack is an int array masked with 0xFFF for 12-bit addresses.

Opcodes:
* execute() method uses ints for data{1,2,3} parameters.
* Store methods force correct masking, eventhough it is strictly needed.

+++ src/main/java/fortville/components/Buffer.java
@@ -7,9 +7,9 @@ import fortville.interfaces.Opcode;
  */
 public class Buffer {
     Opcode opcode;
-    short data1;
-    short data2;
-    short data3;
+    int data1;
+    int data2;
+    int data3;
 
     public Buffer() {
         data1 = 0;
@@ -20,21 +20,21 @@ public class Buffer {
     /**
      * @param data1 the data1 to set
      */
-    public void setData1(short data1) {
+    public void setData1(int data1) {
         this.data1 = data1;
     }
 
     /**
      * @param data2 the data2 to set
      */
-    public void setData2(short data2) {
+    public void setData2(int data2) {
         this.data2 = data2;
     }
 
     /**
      * @param data3 the data3 to set
      */
-    public void setData3(short data3) {
+    public void setData3(int data3) {
         this.data3 = data3;
     }
 
@@ -48,21 +48,21 @@ public class Buffer {
     /**
      * @return the data1
      */
-    public short getData1() {
+    public int getData1() {
         return data1;
     }
 
     /**
      * @return the data2
      */
-    public short getData2() {
+    public int getData2() {
         return data2;
     }
 
     /**
      * @return the data3
      */
-    public short getData3() {
+    public int getData3() {
         return data3;
     }
